refactor: move message manager to its own file
